pyproject.toml
[project]
name = "myhttpserver-test-asset"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "numpy>=2.3.3",
    "opencv-python-headless>=4.12.0.88",
    "pillow>=11.2.1",
]
